Refactor section components. Moved to Detail folder. #400.
