+++ explorer/app/components/Detail/components/Section/components/SectionTitle/sectionTitle.stories.tsx
