+++ explorer/app/components/Detail/components/Section/components/SectionTitle/sectionTitle.tsx
