+++ explorer/app/components/Detail/components/Section/section.styles.ts
