+++ explorer/app/components/Detail/components/Section/section.tsx
