+++ explorer/app/components/Export/components/ExportToTerra/components/ExportToTerraNotStarted/exportToTerraNotStarted.styles.ts
@@ -1,7 +1,7 @@
 // TODO section content and action styles to be refactored with https://github.com/clevercanary/data-browser/issues/400
 import styled from "@emotion/styled";
 import { Stack } from "../../../../../common/Stack/Stack";
-import { SectionActions } from "../../../../../Project/components/Section/section.styles";
+import { SectionActions } from "../../../../../Detail/components/Section/section.styles";
 
 export const SectionContent = styled(Stack)``;
