+++ explorer/app/components/NoResults/noResults.tsx
@@ -1,10 +1,10 @@
 import { Typography } from "@mui/material";
-import { SectionActions } from "app/components/Project/components/Section/section.styles";
+import { SectionActions } from "app/components/Detail/components/Section/section.styles";
 import React, { ReactNode } from "react";
 import { SearchOffIcon } from "../common/CustomIcon/components/SearchOffIcon/searchOffIcon";
 import { RoundedPaper } from "../common/Paper/paper.styles";
 import { PRIORITY, StatusIcon } from "../common/StatusIcon/statusIcon";
-import { SectionTitle } from "../Project/components/Section/components/SectionTitle/sectionTitle";
+import { SectionTitle } from "../Detail/components/Section/components/SectionTitle/sectionTitle";
 import { Section, SectionContent } from "./noResults.styles";
 
 interface Props {
