+++ explorer/app/components/Project/components/Contributors/contributors.tsx
@@ -1,7 +1,7 @@
 import React from "react";
+import { SectionDetailsEmpty } from "../../../Detail/components/Section/components/SectionDetailsEmpty/sectionDetailsEmpty";
+import { Section } from "../../../Detail/components/Section/section";
 import { Contributor } from "../../common/entities";
-import { SectionDetailsEmpty } from "../Section/components/SectionDetailsEmpty/sectionDetailsEmpty";
-import { Section } from "../Section/section";
 import { Sup } from "../Sup/Sup.styles";
 
 interface Props {
