+++ explorer/app/components/index.tsx
@@ -1,6 +1,7 @@
 export { AzulFileDownload } from "./AzulFileDownload/azulFileDownload";
 export { Stack } from "./common/Stack/Stack";
 export { StaticImage } from "./common/StaticImage/staticImage";
+export { Section } from "./Detail/components/Section/section";
 export { ExportMethod } from "./Export/components/ExportMethod/exportMethod";
 export { FileCounts } from "./FileCounts/FileCounts";
 export { IconList } from "./IconList/IconList";
@@ -19,7 +20,6 @@ export { DataReleasePolicy } from "./Project/components/DataReleasePolicy/dataRe
 export { Description } from "./Project/components/Description/description";
 export { Details } from "./Project/components/Details/details";
 export { Publications } from "./Project/components/Publications/publications";
-export { Section } from "./Project/components/Section/section";
 export { SupplementaryLinks } from "./Project/components/SupplementaryLinks/supplementaryLinks";
 export { TitledText } from "./Project/components/TitledText/titledText";
 export { ShowMore } from "./ShowMore/ShowMore";
